+++ examples/nested.rs
@@ -4,12 +4,12 @@ fn main() -> Result<(), Error> {
     let mut taffy = Taffy::new();
 
     // left
-    let child_t1 = taffy.new_node(
+    let child_t1 = taffy.new_with_children(
         Style { size: Size { width: Dimension::Points(5.0), height: Dimension::Points(5.0) }, ..Default::default() },
         &[],
     )?;
 
-    let div1 = taffy.new_node(
+    let div1 = taffy.new_with_children(
         Style {
             size: Size { width: Dimension::Percent(0.5), height: Dimension::Percent(1.0) },
             // justify_content: JustifyContent::Center,
@@ -19,12 +19,12 @@ fn main() -> Result<(), Error> {
     )?;
 
     // right
-    let child_t2 = taffy.new_node(
+    let child_t2 = taffy.new_with_children(
         Style { size: Size { width: Dimension::Points(5.0), height: Dimension::Points(5.0) }, ..Default::default() },
         &[],
     )?;
 
-    let div2 = taffy.new_node(
+    let div2 = taffy.new_with_children(
         Style {
             size: Size { width: Dimension::Percent(0.5), height: Dimension::Percent(1.0) },
             // justify_content: JustifyContent::Center,
@@ -33,7 +33,7 @@ fn main() -> Result<(), Error> {
         &[child_t2],
     )?;
 
-    let container = taffy.new_node(
+    let container = taffy.new_with_children(
         Style { size: Size { width: Dimension::Percent(1.0), height: Dimension::Percent(1.0) }, ..Default::default() },
         &[div1, div2],
     )?;